+++ lib/swid/frontend/dart/mapClassLibrarySourcePath.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/element/element.dart' show Element;
 import 'package:meta/meta.dart';
 
-import 'package:hydro_sdk/swid/ir/util/mapLibrarySourcePath.dart';
+import 'package:hydro_sdk/swid/frontend/dart/mapLibrarySourcePath.dart';
 
 String mapClassLibrarySourcePath({@required Element element}) =>
     mapLibrarySourcePath(librarySource: element?.librarySource);
