+++ lib/swid/frontend/dart/mapCompilationUnitLibrarySourcePath.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/ast/ast.dart' show CompilationUnit;
 import 'package:meta/meta.dart';
 
-import 'package:hydro_sdk/swid/ir/util/mapLibrarySourcePath.dart';
+import 'package:hydro_sdk/swid/frontend/dart/mapLibrarySourcePath.dart';
 
 String mapCompilationUnitSourcePath(
         {@required CompilationUnit compilationUnit}) =>
