+++ lib/swid/frontend/dart/mapLibrarySourcePath.dart
