+++ lib/swid/frontend/dart/swidClassFromInterfaceType.dart
@@ -10,7 +10,7 @@ import 'package:hydro_sdk/swid/ir/swidFunctionType.dart';
 import 'package:hydro_sdk/swid/ir/swidReferenceDeclarationKind.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormal.dart';
 import 'package:hydro_sdk/swid/ir/util/mapAnalyzerNullabilitySuffix.dart';
-import 'package:hydro_sdk/swid/ir/util/mapClassLibrarySourcePath.dart';
+import 'package:hydro_sdk/swid/frontend/dart/mapClassLibrarySourcePath.dart';
 
 import 'package:analyzer/dart/element/element.dart'
     show PropertyAccessorElement;
