+++ lib/swid/frontend/dart/swidFunctionTypeFromFunctionType.dart
@@ -8,7 +8,7 @@ import 'package:hydro_sdk/swid/ir/swidNullabilitySuffix.dart';
 import 'package:hydro_sdk/swid/ir/swidType.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormal.dart';
 import 'package:hydro_sdk/swid/ir/util/mapAnalyzerNullabilitySuffix.dart';
-import 'package:hydro_sdk/swid/ir/util/mapClassLibrarySourcePath.dart';
+import 'package:hydro_sdk/swid/frontend/dart/mapClassLibrarySourcePath.dart';
 import 'package:hydro_sdk/swid/frontend/dart/narrowDartTypeToSwidType.dart';
 
 import 'package:analyzer/dart/element/type.dart'
