+++ lib/swid/frontend/dart/swidInterfaceFromInterface.dart
@@ -5,7 +5,7 @@ import 'package:meta/meta.dart';
 
 import 'package:hydro_sdk/swid/ir/swidReferenceDeclarationKind.dart';
 import 'package:hydro_sdk/swid/ir/util/mapAnalyzerNullabilitySuffix.dart';
-import 'package:hydro_sdk/swid/ir/util/mapClassLibrarySourcePath.dart';
+import 'package:hydro_sdk/swid/frontend/dart/mapClassLibrarySourcePath.dart';
 import 'package:hydro_sdk/swid/frontend/dart/narrowDartTypeToSwidType.dart';
 
 import 'package:analyzer/dart/element/type.dart' show InterfaceType;
